rpl: introduce on_target_del() callback
