+++ 6lbr/rpl/rpl.h
@@ -110,7 +110,7 @@ struct rpl_root {
     bool    rpi_ignorable;
 
     void (*on_target_add)(struct rpl_root *root, struct rpl_target *target);
-    void (*route_del)(struct rpl_root *root, const uint8_t prefix[16], size_t prefix_len);
+    void (*on_target_del)(struct rpl_root *root, struct rpl_target *target);
     void (*on_target_update)(struct rpl_root *root, struct rpl_target *target);
 
     // When enabled, some parts of the specification are ignored in order to
